Also alert on IP/Port scans
Every hour, the top 100 port scans (between a given src and dst) and the
top 100 IP scanners (from a given src) are output, and every single entry
where the number of scanned ports (or IPs) exceeds a threshold result in
an alert.

Notice that the DDoS program as well as the --ddos option has been renamed
to resp. "Security" and --security.